server/requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
mysql-connector-python==8.1.0
python-dotenv==1.0.0
pytest==7.1.2
pytest-mock==3.10.0
Werkzeug==2.3.3